% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_companion_folder.R
\name{add_companion_folder}
\alias{add_companion_folder}
\title{add_companion_folder}
\usage{
add_companion_folder(html_path, src_regex, from_lib, to_lib)
}
\arguments{
\item{html_path}{character. Path to the html file to be edited, serve as the root for lib folder paths}

\item{src_regex}{character. Regex of the existing lib path in the html to replace with new location}

\item{from_lib}{character. Path to the lib folder to copy}

\item{to_lib}{character. Path to copy the lib folder to.}
}
\value{
NULL. 2nd effect : create a lib folder and edit html path accordingly
}
\description{
Copy a lib folder and solve dependencies path in html
}
\examples{
# create tmp dir
temp_dir <- tempfile(pattern = "compile")
dir.create(temp_dir, recursive = TRUE)

# copy html and plugin dir
file.copy(
  from = system.file("template.html", package = "nq1h"),
  to = temp_dir
)

file.copy(
  from = system.file("plugin", package = "nq1h"),
  to = temp_dir,
  recursive = TRUE
)

# run add_companion_folder()
add_companion_folder(
  html_path = file.path(temp_dir, "template.html"),
  src_regex = "__path_to_deps_folder__",
  from_lib = file.path(temp_dir, "plugin"),
  to_lib = file.path(temp_dir, "plugin_copy", "plugin")
)

# clean up
unlink(temp_dir, recursive = TRUE)
}
