% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui.R
\name{build_user_interface}
\alias{build_user_interface}
\title{Layout specification of the user interface (UI)}
\usage{
build_user_interface()
}
\value{
A \code{shiny} user interface definition
}
\author{
Roman Pahl, \email{roman.pahl@gmail.com}
}
\seealso{
\code{\link[shiny]{shinyUI}}
}

