% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_qmd_course.R
\name{compile_qmd_course}
\alias{compile_qmd_course}
\title{Independently compile several qmd and create a common html}
\usage{
compile_qmd_course(vec_qmd_path, output_dir)
}
\arguments{
\item{vec_qmd_path}{character. Vector of the path to qmd files}

\item{output_dir}{character. Output path to store html files and companion folders}
}
\value{
NULL Side effect : generate html files in output dir
}
\description{
Independently compile several qmd and create a common html
}
