% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_html_to_pdf.R
\name{convert_html_to_pdf}
\alias{convert_html_to_pdf}
\title{Convert a html presentation to pdf}
\usage{
convert_html_to_pdf(
  html_path,
  pdf_path,
  pause_before_slide_export = 10,
  debug_dry_run = FALSE
)
}
\arguments{
\item{html_path}{A character. Path to html file.}

\item{pdf_path}{A character. Path to pdf file.}

\item{pause_before_slide_export}{A positive integer. Number of milliseconds to wait before to export each slide.}

\item{debug_dry_run}{A logical. If TRUE the docker command string is returned without running it.}
}
\value{
Nothing. Used for its side effect.
}
\description{
This function is an interface to the dockerised version of \code{decktape}
a cli pdf exporter written in JavaScript.
See  https://github.com/astefanutti/decktape/
}
\details{
It requires that:
\itemize{
\item docker is installed on the system
\item the user running the command is part of the docker group
}
}
\examples{
temp_dir <- tempfile(pattern = "compile")

courses_path <- system.file(
  "courses",
  package = "nq1h"
)
qmds <- list.files(
  path = courses_path,
  full.names = TRUE,
  recursive = TRUE,
  pattern = "qmd$"
)

html_output <- compile_qmd_course(
  vec_qmd_path = qmds,
  output_dir = temp_dir,
  output_html = "complete_course.html"
)

file_path <- list(
  html = file.path(temp_dir, "complete_course.html"),
  pdf = file.path(temp_dir, "complete_course.pdf")
)

convert_html_to_pdf(
  html_path = file_path$html,
  pdf_path = file_path$pdf
)

# browseURL(file_path$pdf)

unlink(temp_dir, TRUE, TRUE)
}
