% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_img_and_edit_path.R
\name{copy_img_and_edit_path}
\alias{copy_img_and_edit_path}
\title{Copy img to root img directory and edit path in html}
\usage{
copy_img_and_edit_path(html_path, img_root_dir)
}
\arguments{
\item{html_path}{character. Path to html file.}

\item{img_root_dir}{character. Path to img root dir}
}
\value{
None. Side effet : edit img path in html.
}
\description{
Copy img to root img directory and edit path in html
}
