\name{corrogram}
\alias{corrogram}
\title{ Draw a color-coded triangular matrix of pairwise correlations }
\description{ This figure is a color-coded, rotated triangular matrix indicating the correlation between every pair of items. }
\usage{
corrogram(...)
}
\arguments{
  \item{\dots}{ Arguments passed to \code{\link{distogram}}. }
}
\details{ 
This is a simple wrapper around \code{\link{distogram}}, with the color scale set by default to use \code{\link{blueorange}} with a range from -1 to +1. 
}
\value{ 
A color map (as generated by \code{\link{makecmap}}), invisibly. 
}
\seealso{ \code{\link{distogram}} }
\examples{ 
  
  corrogram(cor(swiss), title = 'Pearson correlation')
}
\keyword{ hplot }
