% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_html_slides.R
\name{extract_html_slides}
\alias{extract_html_slides}
\title{extract_html_slides}
\usage{
extract_html_slides(vec_html_path)
}
\arguments{
\item{vec_html_path}{character. The vector of path to individual html files}
}
\value{
HTML. The HTML slide content of all html files combined together.
}
\description{
Extract slide content from multiple html
}
\examples{
# list html files
courses_path <- system.file(
  "courses",
  package = "nq1h"
)

htmls <- list.files(
  path = courses_path,
  full.names = TRUE,
  pattern = "html$"
)

html_slide_content <- extract_html_slides(vec_html_path = htmls)
}
