% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\docType{class}
\name{Match-class}
\alias{Match}
\alias{Match-class}
\title{A Squash Match class}
\description{
Models a Squash match consisting of several \code{\link{Game}}s.
 For match play, a "game momentum" can be specified, which increases the odds
 in favor of the player that just has won a game.
}
\section{Fields}{

\describe{
\item{\code{game}}{(Game) contains all necessary game parameters}

\item{\code{bestOf}}{(numeric) the maximum match length}

\item{\code{played_games}}{(list) all played games are saved internally in this list}
}}
\examples{
m1 <- Match(Game())
m1$hasFinished()             # FALSE
m1$get_match_table()         #   1 2 3
                             # A 0 0 0
                             # B 0 0 0
set.seed(123)
m1$play()
m1$hasFinished()             # TRUE
m1$get_match_table()         #    1  2  3  4  5
                             # A 11 11  2  6  9
                             # B  9  2 11 11 11

m1$get_winner()$name         # "B"

}
\author{
Roman Pahl, \email{roman.pahl@gmail.com}
}
\seealso{
\code{\link{Game}},
}

