% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.R
\docType{class}
\name{Player-class}
\alias{Player}
\alias{Player-class}
\title{A Squash Player class}
\description{
Wraps name and skill of a Squash player
}
\section{Fields}{

\describe{
\item{\code{name}}{(character) the player's name}

\item{\code{name}}{(numeric) the player's skill (number between 0 and 1)}
}}
\examples{
p1 <- Player("John", skill = 0.7)
p1$isPlayer()    # TRUE
print(p1)

}
\author{
Roman Pahl, \email{roman.pahl@gmail.com}
}

