% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-server.R
\name{server_function}
\alias{server_function}
\title{Shiny server function}
\usage{
server_function(input, output, session)
}
\arguments{
\item{input}{shiny input parameters}

\item{output}{shiny output parameters}

\item{session}{shiny session parameter}
}
\description{
The server-side logic of the Shiny application is 
defined/started via the function \code{\link[shiny]{shinyServer}}, which 
itself has a function as its argument, which, in turn, is defined here.
}
\details{
Note that this function does not return anything, but rather the
work is done in the function body, using and modifying the \code{input},
\code{output}, and \code{session} parameters as well as calling other shiny
functions such as \code{\link[shiny]{observe}} and
\code{\link[shiny]{observeEvent}}, in order to control elements of the GUI.
}
\author{
Roman Pahl, \email{roman.pahl@gmail.com}
}
\seealso{
For more details see \code{\link[shiny]{shinyServer}}.
}

