% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui.R
\name{set_shiny_theme_if_installed}
\alias{set_shiny_theme_if_installed}
\title{Shiny theme setup}
\usage{
set_shiny_theme_if_installed(theme_name = "united")
}
\arguments{
\item{theme_name}{(character) specifying the theme name}
}
\value{
If the theme is available (i.e., installed) the theme name is
returned as is, else NULL is returned.
}
\description{
Sets a specified shiny theme, but does not fail, if the theme is not
installed. In the latter case, the default theme will be in effect.
}
\author{
Roman Pahl, \email{roman.x.pahl@gsk.com}
}

