% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squash-package.R
\docType{package}
\name{squash-package}
\alias{squash}
\alias{squash-package}
\title{squash: Several Quarto As Single HTML}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Compile n .qmd (under format revealjs) to one single html file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thinkr-open.github.io/squash/}
  \item \url{https://github.com/ThinkR-open/squash}
}

}
\author{
\strong{Maintainer}: Antoine Languillaume \email{antoine@thinkr.fr}

Authors:
\itemize{
  \item Antoine Languillaume \email{antoine@thinkr.fr}
  \item Colin Fay \email{colin@thinkr.fr}
  \item Swann Floc'hlay \email{swann@thinkr.fr}
}

}
\keyword{internal}
