% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Squash.R
\docType{package}
\name{Squash}
\alias{Squash}
\alias{Squash-package}
\title{Squash match simulator}
\usage{
\code{start()}
}
\description{
A package to simulate and visualize Squash matches. At this point, a very
simple probability model is used for this purpose. In particular, a Squash
match consists of two players, for which the winning odds of one player over
the other are specified. Then, a match is simulated point by point, and 
for each point, the winner is decided by sampling from a binomial 
distribution with probability weights specified by the odds. In addition, it
is possible to set a so-called momentum shift, which increases the odds in 
favor of the player that just has won a game.

All functionality is provided in a GUI implemented as a shiny application
where parameters are easily set. The resulting match score in particular is 
provided both as a score table and graphs showing the process of each game.
}
\author{
Roman Pahl, \email{roman.pahl@gmail.com}
}

