% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Squash.R
\name{start}
\alias{start}
\title{Starts the Squash match simulator}
\usage{
start(debugMode = FALSE)
}
\arguments{
\item{debugMode}{(logical) if TRUE, the server function will call
\code{\link[devtools]{load_all()}} each time the browser is refreshed, which
allows to test code-changes without having to abort and restart the function.}
}
\description{
Starts the application in a web browser, which should open automatically.
}
\examples{

\dontrun{
start()
}
}
\author{
Roman Pahl, \email{roman.pahl@gmail.com}
}

