% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_qmd_course.R
\name{compile_qmd_course}
\alias{compile_qmd_course}
\title{compile_qmd_course}
\usage{
compile_qmd_course(
  vec_qmd_path,
  output_dir,
  output_html,
  template = system.file("template.qmd", package = "nq1h")
)
}
\arguments{
\item{vec_qmd_path}{character. Vector of the path to qmd files}

\item{output_dir}{character. Output path to store html files and companion folders}

\item{output_html}{character. File name of the complete html output saved}

\item{template}{character. Path to the template qmd to use. Content will be included at the position where \code{{{ include_html_content }}} is found.}
}
\value{
character. The path to the resulting html file
}
\description{
Independently compile several qmd and create a common html
}
\examples{
# list example qmds
courses_path <- system.file(
  "courses",
  package = "nq1h"
)

qmds <- list.files(
  path = courses_path,
  full.names = TRUE,
  recursive = TRUE,
  pattern = "qmd$"
)

# generate html in temp folder
temp_dir <- tempfile(pattern = "compile")

html_output <- compile_qmd_course(
  vec_qmd_path = qmds,
  output_dir = temp_dir,
  output_html = "course_complete.html"
)

# clean up
unlink(temp_dir, recursive = TRUE)
}
