% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_qmd_course.R
\name{compile_qmd_course}
\alias{compile_qmd_course}
\title{Create a single html from multiple qmd}
\usage{
compile_qmd_course(
  vec_qmd_path,
  output_dir,
  output_html,
  template = system.file("template.qmd", package = "squash"),
  title = "Formation R",
  date = "01/01/01-01/01/01",
  trainer = "ThinkR",
  mail = "thinkr.fr",
  phone = "+33 0 00 00 00 00"
)
}
\arguments{
\item{vec_qmd_path}{character. Vector of the path to qmd files}

\item{output_dir}{character. Output path to store html files and companion folders}

\item{output_html}{character. File name of the complete html output saved}

\item{template}{character. Path to the template qmd to use. Content will be included at the positions inside double-brackets}

\item{title}{character. Title of the presentation}

\item{date}{character. Start and end dates of the training}

\item{trainer}{character. Name of the trainer}

\item{mail}{character. Mail of the trainer}

\item{phone}{character. Phone number of the trainer}
}
\value{
character. The path to the resulting html file
}
\description{
Independently compile several qmd and create a common html
}
\examples{
library(future)

# set parallel rendering
plan(multisession, workers = 2)

# list example qmds
courses_path <- system.file(
  "courses",
  package = "squash"
)

qmds <- list.files(
  path = courses_path,
  full.names = TRUE,
  recursive = TRUE,
  pattern = "qmd$"
)

# generate html in temp folder
temp_dir <- tempfile(pattern = "compile")

html_output <- compile_qmd_course(
  vec_qmd_path = qmds,
  output_dir = temp_dir,
  output_html = "complete_course.html"
)

# reset default rendering
plan("default")

# clean up
unlink(temp_dir, recursive = TRUE)
}
