% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template_html.R
\name{create_template_html}
\alias{create_template_html}
\title{Create a themed html from a qmd template}
\usage{
create_template_html(
  path_to_qmd,
  output_file,
  output_dir,
  title = "Formation R",
  date = "01/01/01-01/01/01",
  temp_dir = tempfile(pattern = "template")
)
}
\arguments{
\item{path_to_qmd}{character. Path to the qmd template to be rendered in thinkridentity-revealjs}

\item{output_file}{character. Name of the output html template}

\item{output_dir}{character. Output path to store html files and companion folders}

\item{title}{character. Title of the presentation}

\item{date}{character. Start and end dates of the training}

\item{temp_dir}{character. Path to the temp_dir where template will be rendered}
}
\value{
character. Path to the html template
}
\description{
Create a template html with ThinkR styling
}
\examples{
# create temp dir
temp_dir <- tempfile(pattern = "template")

# create html template
path_to_html_template <- create_template_html(
  path_to_qmd = system.file("template.qmd", package = "nq1h"),
  output_dir = temp_dir,
  output_file = "complete_course.html"
)

# clean up
unlink(temp_dir, recursive = TRUE)
}
