% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template_html.R
\name{create_template_html}
\alias{create_template_html}
\title{Create a themed html from a qmd template}
\usage{
create_template_html(
  path_to_qmd,
  output_file,
  output_dir,
  output_format = "thinkridentity-revealjs",
  title = "Formation R",
  date = "01/01/01-01/01/01",
  footer = "**<i class='las la-book'></i> Formation R**",
  temp_dir = tempfile(pattern = "template"),
  ext_dir = system.file("_extensions", package = "squash")
)
}
\arguments{
\item{path_to_qmd}{character. Path to the qmd template to be rendered in thinkridentity-revealjs}

\item{output_file}{character. Name of the output html template}

\item{output_dir}{character. Output path to store html files and companion folders}

\item{output_format}{character. Output format of the qmd, default to "revealjs". Can be adapted for specific themes.}

\item{title}{character. Title of the presentation}

\item{date}{character. Start and end dates of the training}

\item{footer}{character. Footer appearing in all slides}

\item{temp_dir}{character. Path to the temp_dir where template will be rendered}

\item{ext_dir}{character. Path to the _extensions directory to use when compiling qmd}
}
\value{
character. Path to the html template
}
\description{
Create a template html with ThinkR styling
}
\examples{
# create temp dir
temp_dir <- tempfile(pattern = "template")

# create html template
path_to_html_template <- create_template_html(
  path_to_qmd = system.file("template.qmd", package = "squash"),
  output_dir = temp_dir,
  output_file = "complete_course.html"
)

# clean up
unlink(temp_dir, recursive = TRUE)
}
