% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_single_qmd.R
\name{render_single_qmd}
\alias{render_single_qmd}
\title{Render a qmd course to html}
\usage{
render_single_qmd(
  qmd,
  img_root_dir = "img",
  output_format = "revealjs",
  metadata = NULL,
  quiet = TRUE,
  purrr_insistently_rate_backoff = purrr::rate_backoff(pause_base = 0.1, max_times = 5)
)
}
\arguments{
\item{qmd}{character. Path to the qmd file to render}

\item{img_root_dir}{character. Path to the main image folder to extract media to}

\item{output_format}{character. Output format of the qmd, default to "revealjs". Can be adapted for specific themes.}

\item{metadata}{list. List of metadata to be used for rendering single qmd file}

\item{quiet}{logical. Output info in user console}

\item{purrr_insistently_rate_backoff}{function. Function to use to retry rendering qmd files in case of failure. Should be a purrr::rate_backoff function.}
}
\value{
logical. TRUE if rendering succeeded, FALSE otherwise. Side effect : render qmd as html
}
\description{
Render a single qmd file to html with image folder.
}
\examples{
# create a temp dir with qmd
temp_dir <- tempfile(pattern = "render")

dir.create(
  path = file.path(temp_dir, "img"),
  recursive = TRUE
)

file.copy(
  from = system.file("courses", "M01", "M01S01", "C01-qmd1_for_test.qmd", package = "squash"),
  to = temp_dir
)

file.copy(
  from = system.file("courses", "M01", "M01S01", "img", "logo_1.png", package = "squash"),
  to = file.path(temp_dir, "img")
)

# render qmd
is_rendered <- render_single_qmd(
  qmd = file.path(temp_dir, "C01-qmd1_for_test.qmd"),
  img_root_dir = file.path(temp_dir, "image_folder")
)

# clean temp dir
unlink(temp_dir, recursive = TRUE)
}
